fixing setup initial company (#181)
* add vscode dev container

* small updates on packages

* update documentation

* remove migrations scripts from source code

* fix errors in setting up initial company
