+++ src/Api/Api.csproj
@@ -18,4 +18,7 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.3" />
   </ItemGroup>
 
+  <ItemGroup>
+      <EmbeddedResource Include="Data\coa.csv" />
+  </ItemGroup>
 </Project>
