+++ src/Api/appsettings.json
@@ -8,10 +8,6 @@
     }
   },
   "Database": {
-<<<<<<< HEAD
     "ConnectionString": "Data Source={0};User ID={1};Password={2};Initial Catalog={3};TrustServerCertificate=yes;MultipleActiveResultSets=true;"
-=======
-    "ConnectionString": "Data Source={0};User ID={1};Password={2};Initial Catalog={3};TrustServerCertificate=yes;"
->>>>>>> main
   }
 }
